Secure endpoints and extend Swagger
