+++ api/gpt4.py
@@ -1,0 +1,1 @@
+from .index import app
